target/arm/kvm-rme: Add breakpoints and watchpoints parameters
Pass the num_bps and num_wps parameters to Realm creation. These
parameters contribute to the initial Realm measurement.

Acked-by: Richard Henderson <richard.henderson@linaro.org>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>

FIXME: num bps and wps provided by KVM. When the VM is not a Realm, they
correspond to the values in the CPU ID register (ID_AA64DFR0_EL1). In a
Realm, they correspond to the values advertised by RMM.