HACK: kvm: Workaround CCA unmap-before-RD issue
Currently the KVM support for CCA reacts badly to changing guest memory
state before a RD is created. Until this is fixed in the Linux branch,
skip the initial memory state change. That's OK to do on CCA because the
guest needs to initialize the RIPAS before accessing RAM, causing a
bunch of returns to QEMU which call set_memory_attributes_private().

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>