hw/vfio/container: Only remove container from address space if it had been added
vfio_container_instance_finalize() may be called on the error path of
vfio_connect_container(), before the container is fully initialized. In
particular if it wasn't added to the parent address space list by
vfio_address_space_insert() yet, calling QLIST_REMOVE() is invalid and
will segfault. Check that the container was added to the address space
before removing it.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>